OPENAI_API_KEY added
