+++ .github/workflows/test.yaml
@@ -79,6 +79,7 @@ jobs:
       AZURE_GPT35_MODEL: ${{ secrets.STAGING_AZURE_GPT35_MODEL }}
       AZURE_OPENAI_API_KEY: ${{ secrets.STAGING_AZURE_OPENAI_API_KEY }}
       TOGETHER_API_KEY: ${{ secrets.STAGING_TOGETHER_API_KEY }}
+      OPENAI_API_KEY: ${{ github.ref_name == 'main' && secrets.PROD_OPENAI_API_KEY || secrets.STAGING_OPENAI_API_KEY }}
     steps:
       - uses: actions/checkout@v4
       - name: Set up Python
@@ -136,6 +137,7 @@ jobs:
       AZURE_GPT35_MODEL: ${{ secrets.STAGING_AZURE_GPT35_MODEL }}
       AZURE_OPENAI_API_KEY: ${{ secrets.STAGING_AZURE_OPENAI_API_KEY }}
       TOGETHER_API_KEY: ${{ secrets.STAGING_TOGETHER_API_KEY }}
+      OPENAI_API_KEY: ${{ github.ref_name == 'main' && secrets.PROD_OPENAI_API_KEY || secrets.STAGING_OPENAI_API_KEY }}
       AUTOGEN_USE_DOCKER: False
     steps:
       - uses: actions/checkout@v4
@@ -165,6 +167,7 @@ jobs:
       AZURE_GPT35_MODEL: ${{ secrets.STAGING_AZURE_GPT35_MODEL }}
       AZURE_OPENAI_API_KEY: ${{ secrets.STAGING_AZURE_OPENAI_API_KEY }}
       TOGETHER_API_KEY: ${{ secrets.STAGING_TOGETHER_API_KEY }}
+      OPENAI_API_KEY: ${{ github.ref_name == 'main' && secrets.PROD_OPENAI_API_KEY || secrets.STAGING_OPENAI_API_KEY }}
       AUTOGEN_USE_DOCKER: False
     steps:
       - uses: actions/checkout@v4
@@ -383,6 +386,8 @@ jobs:
       AZURE_GPT35_MODEL: ${{ github.ref_name == 'main' && secrets.PROD_AZURE_GPT35_MODEL || secrets.STAGING_AZURE_GPT35_MODEL }}
       AZURE_OPENAI_API_KEY: ${{ github.ref_name == 'main' && secrets.PROD_AZURE_OPENAI_API_KEY || secrets.STAGING_AZURE_OPENAI_API_KEY }}
       TOGETHER_API_KEY: ${{ github.ref_name == 'main' && secrets.PROD_TOGETHER_API_KEY || secrets.STAGING_TOGETHER_API_KEY }}
+      OPENAI_API_KEY: ${{ github.ref_name == 'main' && secrets.PROD_OPENAI_API_KEY || secrets.STAGING_OPENAI_API_KEY }}
+
     steps:
       - uses: actions/checkout@v3 # Don't change it to cheackout@v4. V4 is not working with container image.
       # This is to fix GIT not liking owner of the checkout dir - https://github.com/actions/runner/issues/2033#issuecomment-1204205989
