+++ tests/models/llms/test_openai.py
@@ -135,7 +135,7 @@ class TestOpenAI:
         self,
         llm_model: Model,
         api_key_model: Model,
-        llm_config: Dict[str, Any],
+        azure_llm_config: Dict[str, Any],
         user_uuid: str,
         monkeypatch: pytest.MonkeyPatch,
     ) -> None:
@@ -221,7 +221,7 @@ class TestOpenAIAPIKey:
     async def test_openai_api_key_model_create_autogen(
         self,
         api_key_model: Model,
-        llm_config: Dict[str, Any],
+        azure_llm_config: Dict[str, Any],
         user_uuid: str,
         monkeypatch: pytest.MonkeyPatch,
     ) -> None:
