+++ tests/models/teams/test_two_agents_team.py
@@ -231,7 +231,7 @@ class TestTwoAgentTeam:
         enable_monkeypatch: bool,
         llm_model: Model,
         api_key_model: Model,
-        llm_config: Dict[str, Any],
+        azure_llm_config: Dict[str, Any],
         user_uuid: str,
         monkeypatch: pytest.MonkeyPatch,
     ) -> None:
